Added resource service

+++ volunteer-project/src/test/java/edu/mum/cs544/volunteerproject/test/ProjectServiceTest.java
@@ -28,7 +28,9 @@ public class ProjectServiceTest {
 
 	@After
 	public void tearDown() throws Exception {
-		projectService.delete(testProject);
+		if (testProject.getId() > 0) {
+			projectService.delete(testProject);
+		}
 	}
 
 	@Test
